multiple ajax simple idea view fixes

+++ idea-management/src/main/webapp/html/idea/view-list-simple.jsp
@@ -17,34 +17,28 @@
 	import="it.smartcommunitylab.platform.idea.service.IdeaLocalServiceUtil"%>
 <%@ page import="it.smartcommunitylab.platform.idea.model.Idea"%>
 <%@ page import="it.smartcommunitylab.platform.idea.portlet.Utils"%>
+<%@page import="java.util.Random"%>
 
 <%@ include file="/html/common-init.jsp"%>
 
 <%
-	List<Idea> results = (List)request.getAttribute("ideas");
-  Integer currentPage = (Integer)request.getAttribute("_currentPage");
-  Integer delta = (Integer)request.getAttribute("_delta");
-  if (currentPage == null) currentPage = 1;
-  if (delta == null) delta = results.size()+1;
-  
   String baseUrl = (String) request.getAttribute("_baseUrl");
-//   PortletURL portletURL = renderResponse.createRenderURL();
   Long categoryId = ParamUtil.getLong(renderRequest,"categoryId");
   Long callId = ParamUtil.getLong(renderRequest,"callId");
-  Map<String,Object> params = new HashMap<String,Object>();
-  params.put("categoryId", categoryId);
-  params.put("callId", callId);
-  params.put("mvcPath", "/html/idea/view.jsp");
   
-	int offset = delta - results.size();
-	String offsetClass = (offset > 0) ? "offset" + offset*2 : "";
-	java.util.Map<String,String> CC = IdeaLocalServiceUtil.getCategoryColors(scopeGroupId);
 	
 	NumberFormat numberFormat = NumberFormat.getInstance();
 	numberFormat.setMaximumFractionDigits(1);
 	numberFormat.setMinimumFractionDigits(0);
 
 	String listType = GetterUtil.getString(portletPreferences.getValue("listType", Constants.PREF_LISTTYPE_RECENT));
+	boolean pagination = GetterUtil.getBoolean(portletPreferences.getValue("activatePagination", "true"));
+	int delta = 0;
+    if(pagination) {
+    	delta = GetterUtil.getInteger(portletPreferences.getValue("elementInPage", Integer.toString(Constants.PAGINATION_ELEMENTS_IN_PAGE))); 
+    }
+    
+    String confirmMsg =  LanguageUtil.get(locale,"are-you-sure-you-want-to-delete-this");
 %>
 
 <div class="row-fluid">
@@ -53,19 +47,23 @@
 	</span>
 </div>
 
-<div id='result' class="idea-slider row-fluid">
+<div id="<portlet:namespace/>result" class="idea-slider row-fluid">
+</div>
 
 <aui:script use="liferay-portlet-url">
 	var url = Liferay.PortletURL.createResourceURL();
-	url.setParameter("cur", "1");
     url.setPortletId('<%= Constants.IDEA_PORTLET_ID %>');
-	url.setParameter("categoryId", <%= Long.toString(categoryId)%>);
-	url.setParameter("callId", <%= Long.toString(callId)%>);
-	paginateIdeas(url.toString());
+	url.setParameter("cur", "1");
+    url.setParameter("listType",'<%= listType %>');
+    url.setParameter("pagination",'<%= Boolean.toString(pagination) %>');
+    url.setParameter("delta",'<%= Integer.toString(delta) %>');
+	url.setParameter("categoryId", '<%= Long.toString(categoryId) %>');
+	url.setParameter("callId", '<%= Long.toString(callId) %>');
+	<portlet:namespace/>paginateIdeas(url.toString());
 </aui:script>
 
 	<aui:script>
-		Liferay.provide(window, 'paginateIdeas', function(url) {
+		Liferay.provide(window, '<portlet:namespace/>paginateIdeas', function(url) {
 			var instance = this;
 			var A = AUI();
             //alert(url);
@@ -75,9 +73,8 @@
 					success : function() {
 						var data = this.get('responseData');
                         // transform data
-                        if( data.currentPage == 1) {
-                        	data.currentPage = 0; // for handlebars templata
-                        }
+                        data.viewPrevArrow = <%= pagination %> && data.currentPage != 1;
+                        data.viewNextArrow = <%= pagination %>;
                         
                         var offset = data.elementInPage - data.result.size;
                         data.offsetClass = (offset > 0) ? "offset" + offset*2 : "";
@@ -91,11 +88,12 @@
                                 v.stars[i-1] = i <= v.avgRating;
                             }
                         });
-                        
-						var nodeResult = A.one('#result');
+                        //alert(data.result.size);
+						var nodeResult = A.one('#<portlet:namespace/>result');
+                        //alert(nodeResult);
 						nodeResult.empty();
 						var handlebars = new A.Template(A.Handlebars);
-						var source = A.one("#idea-templ").html();
+						var source = A.one('#<portlet:namespace/>idea-templ').html();
 						var compiled = handlebars.compile(source);
 						var html = compiled({
 							result : data
@@ -107,25 +105,23 @@
                 });
             }, [ 'aui-io', 'aui-node', 'template-base', 'handlebars' ]);
 	</aui:script>
-	
 
 
-	
-<script id="idea-templ" type="text/x-handlebars">
+<script id="<portlet:namespace/>idea-templ" type="text/x-handlebars">
     <span class="span1 text-right"> 
-		{{#if result.currentPage }}
-			<a onclick="paginateIdeas('{{result.prevURL}}');"> <i
+		{{#if result.viewPrevArrow }}
+			<a onclick="<portlet:namespace/>paginateIdeas('{{result.prevURL}}');"> <i
 				class="icon-arrow-left idea-slider-arrow"></i>
 			</a>
 		{{/if}}
 	</span>
     {{#each result.result.data}}
      <span class="span2" id="result">
-   <div id="resContainer" onClick="javascript:window.location = {{this.detailURL}};" class="thumbnail" style="border-left-color: '{{this.categoryColor}}';">
+   <div id="<portlet:namespace/>resContainer" onClick="javascript:window.location = {{this.detailURL}};" class="thumbnail" style="border-left-color: {{this.categoryColor}};">
 	<div class="idea-cat">
 				<span style="color: {{this.categoryColor}};">{{this.category}}</span>
 				{{#if this.deleteURL}}
-					<a id="delete-link-{{@index}}" href="{{this.deleteURL}}" class="pull-right"><span class="delete-icon" style="height:16px;width:16px;display:inline-block;"></span></a>
+					<a id="delete-link-{{@index}}" href="{{this.deleteURL}}" onclick="return confirm('<%= confirmMsg %>')" class="pull-right"><span class="delete-icon" style="height:16px;width:16px;display:inline-block;"></span></a>
 				{{/if}}			
 			</div>
                     <h4>{{this.title}}</h4>
@@ -148,10 +144,10 @@
     {{/each}}
 	
 	<span class="span1 text-left {{result.offsetClass}}"> 
-			<a onclick="paginateIdeas('{{result.nextURL}}');"> <i
+		{{#if result.viewNextArrow }}
+			<a onclick="<portlet:namespace/>paginateIdeas('{{result.nextURL}}');"> <i
 				class="icon-arrow-right idea-slider-arrow"></i>
 			</a>
+		{{/if}}
 	</span>
 </script>
-
-</div>
